fix: SidebarView type-checker timeout, #Preview macros, iOS #if os(iOS) guards
- SidebarView.swift: Extracted ForEach row into cycleListRow() helper to fix
  Swift type-checker timeout at line 156
- Removed duplicate OKRCycle definition (now uses shared library's OKRCycle)
- Replaced .accessibilitySelected with .accessibilityAddTraits for macOS compat
- Wrapped #Preview macros in #if !SWIFT_PACKAGE for SPM CLI builds
  (SidebarView, MacTreeView, OKRAlignmentMacApp)
- Wrapped all iOS target files with #if os(iOS) guards
  (OKRAlignmentApp, iOSTreeView, iOSNodeEditSheet, iOSNodeDetailView, TabBarView)
- EntityMappingTests and ViewModelTests rewritten to test production code
  using in-memory CoreData stack (CoreDataTestStack)
- Shared library and Mac target both compile cleanly

+++ project/OKRAlignment/Sources/OKRAlignmentShared/Data/Mappers/DomainToEntityMapper.swift
@@ -72,7 +72,7 @@ enum DomainToEntityMapper {
         entity.ownerName = domain.ownerName
         entity.createdAt = domain.createdAt
         entity.updatedAt = domain.updatedAt
-        entity.sortOrder = Int32(domain.sortOrder)
+        entity.sortOrder = Int64(domain.sortOrder)
         
         // 关联周期（如果指定了cycleId）
         if let cycleId = domain.cycleId {
